move pubsub test utils to tools
