+++ libp2p/tools/constants.py
@@ -13,7 +13,6 @@ MAX_READ_LEN = 2 ** 32 - 1
 LISTEN_MADDR = multiaddr.Multiaddr("/ip4/127.0.0.1/tcp/0")
 
 
-
 FLOODSUB_PROTOCOL_ID = floodsub.PROTOCOL_ID
 GOSSIPSUB_PROTOCOL_ID = gossipsub.PROTOCOL_ID
