+++ tests/pubsub/test_floodsub.py
@@ -4,12 +4,11 @@ import pytest
 
 from libp2p.peer.id import ID
 from libp2p.tools.factories import FloodsubFactory
-from libp2p.tools.utils import connect
-
-from .floodsub_integration_test_settings import (
+from libp2p.tools.pubsub.floodsub_integration_test_settings import (
     floodsub_protocol_pytest_params,
     perform_test_from_obj,
 )
+from libp2p.tools.utils import connect
 
 
 @pytest.mark.parametrize("num_hosts", (2,))
