+++ tests/pubsub/test_gossipsub_backward_compatibility.py
@@ -4,8 +4,7 @@ import pytest
 
 from libp2p.tools.constants import FLOODSUB_PROTOCOL_ID
 from libp2p.tools.factories import GossipsubFactory
-
-from .floodsub_integration_test_settings import (
+from libp2p.tools.pubsub.floodsub_integration_test_settings import (
     floodsub_protocol_pytest_params,
     perform_test_from_obj,
 )
